+++ yggdrasil_decision_forests/port/python/tools/build_linux_aarch64_release.sh
@@ -16,7 +16,7 @@
 
 # Builds all python versions for release on Pypi
 
-PYTHON_VERSIONS=( 3.8 3.9 3.10 3.11 )
+PYTHON_VERSIONS=( 3.8 3.9 3.10 3.11 3.12 )
 
 function build_py() {
   local PYTHON="python"$1