+++ yggdrasil_decision_forests/port/python/tools/build_linux_release.sh
@@ -16,11 +16,11 @@
 
 # Builds all python versions for release on Pypi
 
-PYTHON_VERSIONS=( 3.8 3.9 3.10 3.11 )
+PYTHON_VERSIONS=( 3.8 3.9 3.10 3.11 3.12 )
 
 function build_py() {
   local PYTHON="python"$1
-  echo "Starting build with " $PYTHON
+  echo "Starting build with $PYTHON"
   $PYTHON -m venv /tmp/venv_$PYTHON
   source /tmp/venv_$PYTHON/bin/activate
   bazel clean --expunge