+++ yggdrasil_decision_forests/port/python/ydf/learner/learner_test.py
@@ -496,15 +496,15 @@ class RandomForestLearnerTest(LearnerTest):
 
   def test_default_hp_dictionary(self):
     learner = specialized_learners.RandomForestLearner(label="l", num_trees=50)
-    self.assertDictContainsSubset(
+    self.assertLessEqual(
         {
             "num_trees": 50,
             "categorical_algorithm": "CART",
             "categorical_set_split_greedy_sampling": 0.1,
             "compute_oob_performances": True,
             "compute_oob_variable_importances": False,
-        },
-        learner.hyperparameters,
+        }.items(),
+        learner.hyperparameters.items(),
     )
 
   def test_multidimensional_training_dataset(self):